add the 0 file

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -12,7 +12,7 @@
 int sum_them_all(const unsigned int n, ...)
 {
 	va_list valist;
-	double sum = 0.0;
+	int sum = 0;
 	unsigned int i;
 
 	if (n == 0)
